+++ auth/domain/src/main/kotlin/me/proton/core/auth/domain/usecase/SetupAccountCheck.kt
@@ -30,8 +30,6 @@ import me.proton.core.domain.entity.UserId
 import me.proton.core.user.domain.extension.hasInternalAddressKey
 import me.proton.core.user.domain.extension.hasKeys
 import me.proton.core.user.domain.extension.hasUsername
-import me.proton.core.user.domain.extension.isPrivate
-import me.proton.core.user.domain.extension.isSubUser
 import me.proton.core.user.domain.repository.UserAddressRepository
 import me.proton.core.user.domain.repository.UserRepository
 import javax.inject.Inject
@@ -65,7 +63,8 @@ class SetupAccountCheck @Inject constructor(
     suspend operator fun invoke(
         userId: UserId,
         isTwoPassModeNeeded: Boolean,
-        requiredAccountType: AccountType
+        requiredAccountType: AccountType,
+        isTemporaryPassword: Boolean
     ): Result {
         val user = userRepository.getUser(userId, refresh = true)
         return when (requiredAccountType) {
@@ -84,7 +83,7 @@ class SetupAccountCheck @Inject constructor(
                 val addresses = addressRepository.getAddresses(userId, refresh = true)
                 when {
                     !user.hasUsername() -> ChooseUsernameNeeded
-                    !user.hasKeys() && user.isSubUser() && user.isPrivate() -> ChangePasswordNeeded
+                    isTemporaryPassword -> ChangePasswordNeeded
                     !user.hasKeys() -> SetupPrimaryKeysNeeded
                     !addresses.hasInternalAddressKey() -> SetupInternalAddressNeeded
                     isTwoPassModeNeeded -> TwoPassNeeded