+++ auth/domain/src/test/kotlin/me/proton/core/auth/domain/usecase/CreateLoginSessionTest.kt
@@ -112,7 +112,8 @@ class CreateLoginSessionTest {
                 "server-proof",
                 1,
                 0,
-                secondFactor
+                secondFactor,
+                false
             )
         }